Add matrix wrapper classes for operators
AbstractMatrixWrapper wraps objects of type AbstractHamiltonian and
AbstractObservable (both representing operators acting on a Hilbert
space) and provides a method to apply these operators to state vectors.

SparseMatrixWrapper and DenseMatrixWrapper extract the matrix elements
from the operator and stores them as an Eigen sparse or dense matrix,
respectively. This is of course only feasible for small Hilbert space
dimensions.

The dense and sparse wrappers also provides a method to obtain the
eigendecomposition of the wrapped operator.

DirectMatrixWrapper does not store the matrix elements separately but
iterates through all basis states of the Hilbert space and calls
FindConn every time Apply is called.