Add basic ODE time stepping
This commit adds code for ODE time propagation using two
simple explicit steppers (using the Euler and RK4 method).

Both included steppers are mainly for demonstration purposes.
For productive use, adaptive step-size methods with error control
are more appropriate and will be added later.

The unit tests will use boost::odeint for comparison if it is found on
the system.