Small optimizations for controlled time steppers
In this Dopri54 implementation, the last k_ value is the same as the
first for the next time step (this is called FSAL, first same as last,
in the literature), so it is now reused when the last step was not
rejected.

Also, ControlledTimeStepper now stores the time step between different
calls to Propagate.