Allow duplicate connections when initializing matrix wrappers
The matrix wrapper implementation assumed that each pair of indices
(i,j) occurs only once when collecting the results of
Hamiltonian::FindConn over all states using HilbertIndex.

This is not true in some cases, e.g., for a 2x2 cubic lattice with
periodic boundary conditions. With this commit multiple matrix
elements for the same edges are now added together.