Speed-up calculation of the matrix elements
This commit speeds up the computation of matrix elements previously done
in Hamiltonian::FindConn by providing a new method ForEachConn that
iterates over all reachable configurations calling a callback function
each time. This removes the overhead from std::vector insertion
operations that is present in FindConn.

For backwards compatibility, ForEachConn is implemented in
AbstractHamiltonian in terms of FindConn, but may be overriden by
subclasses. This is currently only done for Ising, which shows a
noticeable speedup for larger systems.