Fix segmentation fault in machine test
The reason for the segfault seems to have been that the Machine
instances only store references to the Hilbert instances.
Since these were created from Python and stored in the variable `hi`,
they were deleted when `hi` was assigned another Hilbert instance
in the test code and the machines then contained dangling references.