Fix warning due to string comparison with `is`
Strings should be compared with `==`. Comparing with `is` checks for
object identity, which works accidentally here but is not guaranteed to [1],
because equal strings are not necessarily the same object. For this reason,
the code produces warnings in Python 3.8.

Further, `name is "mean" or "Mean"` evaluates to `True` if `name is
"mean"` but to "Mean" otherwise because it is parsed as
```
	(name is "mean") or "Mean"
```
which, while not breaking the code, is also unexpected.

In contrast, checking for membership with `name in (...)` also checks
for equality (`==`).

[1] https://docs.python.org/3.8/whatsnew/3.8.html#changes-in-python-behavior