Rename implementation files in netket.operator
As brought up in issue #445, it can be confusing (at least with
auto-completion) if `netket.operator` contains both `LocalOperator`
directly (which users are supposed to use) and the submodule
`local_operator.py` where that class is implemented. Thus, this
implementation file is renamed. (And so are all the others, to which the
same argument applies.)