Remove `EdResult` and streamline ED interface.
The existence of the `EdResult` class was purely for reasons of
backwards compatibility with the old C++ interface. The step to version
3.0 provides the ideal oportunity to remove it.

Directly returning the array of eigenvalues and, optionally,
eigenvectors is in line with Numpy and Scipy and makes the NetKet
interface more comfortable to use. (It is just a thin convenience
wrapper anyways.)