Remove `EdResult` and streamline ED interface (#511)
* Return ndarray from AbstractOperator.to_dense, instead of numpy.matrix

The use of `numpy.matrix` is no longer recommended according to the
official documentation and can lead to surprising errors when used in
ocde that expects to deal with `ndarray`s.

* Remove `EdResult` and streamline ED interface.

The existence of the `EdResult` class was purely for reasons of
backwards compatibility with the old C++ interface. The step to version
3.0 provides the ideal oportunity to remove it.

Directly returning the array of eigenvalues and, optionally,
eigenvectors is in line with Numpy and Scipy and makes the NetKet
interface more comfortable to use. (It is just a thin convenience
wrapper anyways.)

* Update ED unit tests

* Fix docstring result

* Fix Heisenberg test and AKLT_ed example

* Test scipy_args argument of lanczos_ed

* Use correct return type in to_dense docs