Use __matmul__ (@) for the operator product
In 2014, the Python community has adopted PEP 465 which introduced a
__matmul__ (@) operator in order to distinguish it from __mul__ (*) which is
commonly used for elementwise multiplication.

Since this is a widely adopted convention in Python (in particular numpy),
it makes sense for us to follow it as well. Then, expressions such as

id = -1j * sx @ sy @ sz

would have the same semantics for NetKet operators as for numpy arrays.