Add support for coloured graphs to `Heisenberg` (#972)
This PR allows to use Heisenberg operators on coloured graphs, using a different J and sign-rule policy on edges with different colours, so it can be used with #970 easily.

The original calling sequence is still valid. In addition, J can be a list with as many elements as colours in the graph: the first element is used for colour 0, etc. By default, there are no sign rules, but a sign_rule list can be passed (must have the same length as J).