[NFC] Cleanup MCState and reorder the expect code (#954)
This PR moves around the files of MCState/MCMixedState to disentangle them a bit.
Also simplifies how expect is written for most operators.

For the forward pass (and only for the forward pass) we can get by with only one dispatch, as you wanted. However, I'm still keeping the double dispatch around because 1) it's needed for chunking and 2) it's handy for hacking on netket.