Move `sample_next` to `MetropolisSampler` (#1016)
After discussion, we've decided that n_chains should not be completely removed in exact samplers, because we want a unified interface for all samplers, and n_chains can be used in PT and pmap samplers.

Now I'm going to break #1013 down to several smaller PRs.

API changes:

The method sample_next in Sampler and exact samplers (ExactSampler and ARDirectSampler) is removed, and it is only defined in MetropolisSampler. The module function nk.sampler.sample_next also only works with MetropolisSampler. For exact samplers, please use the method sample instead.