Fix factor of 2 on gradients of models with real parameters (#1069)
Multiply by 2 the gradient (`use_covariance=True`) of models with real parameters. The previous value was wrong, because when we take the real part we don't multiply by 2. See Eq.3.15 of https://gitlab.com/nqs/ictp_school/-/blob/master/notes.pdf

THIS IS BREAKING because some models will now have a gradient doubled (so learning rate must be changed) but I think it's a bug fix