Support Python 3.10 (#952)
With this PR (in particular, the change by @inailuig ) NetKet now runs on Python 3.10!

Instead of adding a new tester, which I think would be quite heavy, I just changed the python 3.9 tester to python 3.10.
We don't test throughly 3.9 but... that's ok (we test 3.7, 3.8 and 3.10). And we run the doctests on 3.9 so it's still roughly tested.