Speedup continuous hilbert sampling (and Particle hilbert fix) (#1083)
There are two commits in this PR. The first simply removes the _L attribute that was duplicated between Particle and ContinuouHilbert. That not important

The latter speeds up gaussian sampling by using np.* instead of jnp.* on all calculations that are carried out on constants.
Since PBC/extent are compile-time constants, by using numpy we perform those calculation at compile time and the resulting code will be slightly faster.

Almost negligible, but still, that's a better way to write it down.