Implementation of fermionic hilbert and operator (#1090)
This is an implementation of discrete fermionic degrees of freedom by @imi-hub and myself.

It contains the hilbert space and an operator. The operator handles the minus signs coming from the exchange symmetry, while the hilbert space stores occupation numbers only. This follows e.g. the way qiskit and openfermion would implement things.

For the operator, we follow the implementation of openfermion and added an 'from_openfermion' function to create operators. This allows us to use all their hamiltonians, such as Fermi-Hubbard.

Co-authored-by: Imelda Romero <imelda.romero@epfl.ch>
Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>