Support Windows (#1118)
To support Windows, the most nasty thing is that the default size of some int is 32 bits rather than 64 bits, even on 64-bit Windows. (I didn't test it on 32-bit Windows, but come on it's 2022)

Also, we need to correctly find mpicc because the default configuration of mpi4py on Windows does not provide it.