Deprecate .dtype from all layers and modules, switch to .param_dtype (#1214)
This is to be consistent with the new flax. It's annoying, but it's for the better.

It will be mildly annoying for users by trying to be loud, but we don't break code because everything is properly deprecated.

I kind-of-like it as I prefer RBMModPhase(param_dtype=float32)to RBMModPhase(dtype=float32) to use single-precision.