One arm irreps (#1260)
One of the observations I had with GCNNs was that multidimensional irreps are usually more stable when you project on a specific subspace rather than allowing the full irrep space. A common scenario is when two different wave vectors are forced to be degenerate by a point group symmetry that maps them on one another. This PR introduces one_arm_irreps as an alternative to space_group_irreps which only returns the component of the irrep with the given wave vector, not all the others in its star. Strictly speaking these are not irrep characters, but can be used in GCNN projections equally well (you just project on a narrower subspace).

(Footnote: "arms" of "stars" are standard crystallographic terminology for wave vectors related by point group symmetries.)