implement nn.binary_encoding (#1209)
This way of converting configurations of arbitrary local size Hilbert spaces is compact, as it does not use one-hot encoding (as RBMMultiVal does). Having this compact representation has the cost that the NDM needs to learn the correlations between all the bits that correspond to the same occupation number. In the example above [1, 1, 0] is split into [1], which represents the state of the two-level system, and [1,0], which represents the state of the three-level system.

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>