allow option to disable jitting of ode solvers in TDVP (#1304)
In most things I'm doing nowadays, I cannot jit the TDVP solvers.
Moreover, not sitting them leads to huge speedups in compile time.

This PR introduces a new configuration flag that allows us to turn it off. My plan would be to disable the hitting by default in the next release if this works well.