Add `LogStateVector` model that stores the log- of the wavefunction (#1324)
With the recent release of jax 0.3.17 we can now officially use callbacks inside of our models. 

It's exponentially costly, but it is handy sometimes to check if the problem is the sampling or something else. Also, in combination with `ExactState` (which in hindsight should have been called `FullSummationState`), it allows to have a 100% exact state in our calculations.

Co-authored-by: Damian Hofmann <femtobit@users.noreply.github.com>