use complex-aware activations as default for arnn (#1371)
Up to now we had an activation that was not doing nothing on the
imaginary part of the input. This PR uses the `reim_` version. It does
not change behaviour for real parameters, it does change the behaviour
for complex but the old behaviour was clearly wrong so this is in a
sense a bug fix