Add VMC from scratch tutorial (#1586)
I gave this in trento. It explains how to build a vmc code from scratch
without a variational state, but only using netket operators and
samplers.
It's a great entry point for students.
I also have one to implement tVMC, that I'll eventually clean up and add
as well.

Maybe I should clean it up a bit and remove all the introductory blabla
on operators and mean field?

There's a few magic tricks I pull off, so one should look at the preview
on readthedocs rather than the notebook to see how it is rendered.