SpinOrbitalFermion simplifications, add Slater model, simplify G. tutorial (#1622)
This PR does three things,
- 1 It proposes a way to address #1606 by changing the API of
`SpinOrbitalFermions` to be easier to use. Notably:
- `hi.n_fermions` is always an integer and returns the total number of
fermions in the Hilbert space if specified, `None` otherwise. This is
different from current behaviour where it returns `None` or an integer
only for spineless fermions, and a tuple of those values if they are
spinful.
- Add a new `hi.n_fermions_per_spin` (accepting better naming proposals)
that returns a `tuple[Optional[int], ...]` that is, a tuple of `int` or
`None` if there is a set number of fermions in that spin subspace. If
the fermions are spinless it still returns a tuple with a single element
for consistency and to simplify user code, that this way does not have
to check multiple types/iterable/non iterable
- (It also cleans a bit up the API and ensures users cannot modify the
Hilbert space number of fermions/orbitals because that's unsupported and
they have to construct a new one)
   - I think this makes code easier.

- 2 Implements a `Slater` ansatz for spinless/spinful fermions in
`nk.models`. Interestingly, the code is easier to read thanks to the new
interface, otherwise we'd have to insert many checks.
- 3 Changes the tutorial by gcarleo by adding a few extra comments but
mainly swapping the dimensions of the M matrix from `NFermions x
NOrbitals` to `NOrbitals x NFermions` . This makes the code easier to
read and marginally faster because we're indexing the 'correct'
dimension.