Add RNN (#1305)
RNNs are a subclass of ARNNs, and they reuse ARNN's sampler. They
support complex parameters using the same normalization as ARNNs, and
I'll leave the mod-phase version to a more general implementation of
`ModPhase`.

When using an RNN on a 2D (or higher-D) lattice, we usually need a
non-trivial autoregressive order (like the snake order) to utilize the
locality of the lattice. For 1D RNN it's straightforward to implement
because the RNN cell only needs to access the previous site in the
autoregressive order, without knowing the graph. However, for 2D RNN it
can be complicated because the RNN cell needs to access multiple
previous neighbors, which are defined by both the order ('previous') and
the graph ('neighbor').

The methods `reorder` and `inverse_reorder` have been added to
`AbstractARNN` because `ARDirectSampler` needs them, but currently there
is no way to specify ordering for existing ARNNs.

There is a helper function `netket.nn.rnn.get_snake_inv_reorder_idx` to
generate the snake order for a 2D graph, and `_get_inv_reorder_idx` as a
last sort for an arbitrary graph, which picks a same number of previous
neighbors at every site (if not... that's the job of graph RNN).

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>