Support matmul between `DiscreteOperator` and sparse array (#1661)
So I'm going to make a bunch of PRs for small quality-of-life
improvements when I was playing with RNN...

By default `DiscreteOperator.__matmul__` converts the operator to a CSR
matrix, so it should be straightforward to multiply it onto a sparse
array (usually but not necessarily a vector). For best performance the
sparse array should be in CSR format.

Currently in jax there is no implicit conversion between jax sparse
arrays and scipy sparse arrays, so we follow this limitation and
`DiscreteOperator` only supports scipy sparse arrays, while
`DiscreteJaxOperator` only supports jax sparse arrays. The most well
supported jax sparse array format is BCOO currently.