Override `to_local_operator` to return `LocalOperatorJax` instead of `LocalOperator` (#1694)
When rescaling and/or shifting `IsingJax` by numbers, the operator is
casted to a `LocalOperator`, while it is desirable to keep it as a
`LocalOperatorJax` (for instance if you want to sample from H|psi> where
H is the rescaled and/or shifted `IsingJax` Hamiltonian and |psi> an
arbitrary state, you need H to be `LocalOperatorJax`). In this PR, the
function `to_local_operator` inside `IsingJax` is overridden to achieve
this.