Finalize deprecations from > 2 years ago (#1698)
* Finalize deprecations of some old methods in `nk.sampler` namespace
(see original commit
[1f77ad8267e16fe8b2b2641d1d48a0e7ae94832e](https://github.com/netket/netket/commit/1f77ad8267e16fe8b2b2641d1d48a0e7ae94832e))
- Original deprecation in December 2021, commit
https://github.com/netket/netket/commit/1f77ad8267e16fe8b2b2641d1d48a0e7ae94832e
* Finalize deprecations of 2D input to DenseSymm layers, which now turn
into error and `extra_bias` option of Equivariant Networks/GCNNs (see
original commit
[c61ea542e9d0f3e899d87a7471dea96d4f6b152d](https://github.com/netket/netket/commit/c61ea542e9d0f3e899d87a7471dea96d4f6b152d))
- Original deprecation in December 2021, commit
https://github.com/netket/netket/commit/c61ea542e9d0f3e899d87a7471dea96d4f6b152d
* Finalize deprecations of very old input/properties to Lattices
[0f6f520da9cb6afcd2361dd6fd029e7ad6a2693e](https://github.com/netket/netket/commit/0f6f520da9cb6afcd2361dd6fd029e7ad6a2693e))
- original deprecation in May 2020 commit
https://github.com/netket/netket/commit/0f6f520da9cb6afcd2361dd6fd029e7ad6a2693e


This removes support for stuff we deprecated up to netket v3.3, released
in December 2021.

The next round of deprecations are those from netket v3.5, released in
August 2022, and are the `nk.nn.Module` and `nk.nn.Dense` which at that
point we removed from netket because it was the first release of flax to
support complex number. Also, we switched from `dtype` to `param_dtpye`
as flax. I prefer to keep those around for 2 full years (so to be
removed in august 2024) as those things ended up in several
notebooks/tutorials..